requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm @ file:///croot/comm_1709322850197/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1731428322366/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Django==5.1.6
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
flatbuffers==25.1.24
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1738940303262/work
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1734548052611/work
jedi @ file:///croot/jedi_1733987392413/work
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
keras==3.8.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459213453/work
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
ml-dtypes==0.4.1
munkres==1.1.4
namex==0.0.8
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
opt_einsum==3.4.0
optree==0.14.0
packaging @ file:///croot/packaging_1734472117206/work
parso @ file:///croot/parso_1733963305961/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929693232/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.6
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1735698276679/work
PySide6==6.8.2
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pyzmq @ file:///croot/pyzmq_1734687138743/work
requests==2.32.3
rich==13.9.4
shiboken6==6.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.5.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tornado @ file:///croot/tornado_1733960490606/work
traitlets @ file:///croot/traitlets_1718227057033/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692496989/work
urllib3==2.3.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==3.1.3
wrapt==1.17.2
